refct: rename facetlist package
