+++ instantsearch-compose/src/main/java/com/algolia/instantsearch/compose/filter/facet/FacetListConnector.kt
@@ -1,6 +1,6 @@
-package com.algolia.instantsearch.compose.filter.list
+package com.algolia.instantsearch.compose.filter.facet
 
-import com.algolia.instantsearch.compose.filter.list.internal.FacetListConnectionPager
+import com.algolia.instantsearch.compose.filter.facet.internal.FacetListConnectionPager
 import com.algolia.instantsearch.compose.paging.Paginator
 import com.algolia.instantsearch.core.connection.Connection
 import com.algolia.instantsearch.helper.filter.facet.FacetListConnector
