+++ instantsearch-compose/src/main/java/com/algolia/instantsearch/compose/filter/facet/FacetListState.kt
@@ -1,6 +1,6 @@
-package com.algolia.instantsearch.compose.filter.list
+package com.algolia.instantsearch.compose.filter.facet
 
-import com.algolia.instantsearch.compose.filter.list.internal.FacetListStateImpl
+import com.algolia.instantsearch.compose.filter.facet.internal.FacetListStateImpl
 import com.algolia.instantsearch.core.selectable.list.SelectableItem
 import com.algolia.instantsearch.helper.filter.facet.FacetListView
 import com.algolia.search.model.search.Facet
