+++ instantsearch-compose/src/main/java/com/algolia/instantsearch/compose/filter/facet/internal/FacetListStateImpl.kt
@@ -1,9 +1,9 @@
-package com.algolia.instantsearch.compose.filter.list.internal
+package com.algolia.instantsearch.compose.filter.facet.internal
 
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.setValue
-import com.algolia.instantsearch.compose.filter.list.FacetListState
+import com.algolia.instantsearch.compose.filter.facet.FacetListState
 import com.algolia.instantsearch.core.Callback
 import com.algolia.instantsearch.core.selectable.list.SelectableItem
 import com.algolia.search.model.search.Facet
